This does two things. Most notably, it adds MySQL to SecureDrop. This is necessary for the generation of source designations by the journalist. HACKING.md and setup_ubuntu.sh have both been modified to add this piece of infrustructure. setup_ubuntu requires no additional user input for this step, and modifies config.py accordingly.
This adds an additional 2 pip dependencies:
MySQL-python
SQLAlchemy

The db.py module uses SQLAlchemy to define the tables and perform
database functions.

We are storing the codenames in the database which can be queried with
the sid.

New codenames are generated with two new dictionaries, located in
deaddrop/dictionaries, with an adjective and noun pair. This is
memorable for the journalist. These dictionaries were generated with the
python natural language toolkit, using the wordnik database.