remove ref to deaddrop, and fake email from uid
Since the reference to deaddrop was part of the fake email address, it made
sense to remove them both since the fake email was not actually needed.
I modified getkey so it should still work, even if used with reply keypairs
generated with previous versions of securedrop (they'll still have the
hashed codename in their uid).