don't use virtualenv for dev setup
virtualenv doesn't mix well with Vagrant. By storing the virtualenv inside the
shared folder, we were unintentionally "caching" the Python dependencies
between cycles of vagrant destroy/up. This was accidentally good for
performance, but is counter to the reproducible philosophy of Vagrant and
sometimes lead to confusing problems.

This also simplifies the development cycle because you no longer need to
remember to activate the virtual environment.

Now, setup_dev.sh installs the Python dependencies globally with `sudo pip
install`.