Consolidate configuration into a single config file
Simplifies and consolidates configuration into a single file, which
makes it easier to reason about for developers and work with for
deployment. Updates other code accordingly.

Changes secret keys so we now use a different SECRET_KEY (used by Flask
for signing cookies) for the source and document interfaces. This
exposes some interesting issues with cookies in development.