Logo links to lookup view if source is logged in
Currently, the logo always links to the index. This is a little bit
different from the standard web UX, where such a "home" link takes a
logged in user to their user-specific home page. I'm concerned that the
current behavior may lead to users being locked out of the account they
just created if they accidentally click it without having memorized or
recorded their codename.

This change checks if the user is logged in, and if so, makes the logo
link back to their lookup page, where they can make submissions and
receive replies.