Remove request_that_secures_file_uploads
This was an example integration for a proposed mitigation for the
"plaintext remains in memory" issue (#99), from when we were trying to
solve it by replacing the underlying stream class used by Flask for
requests. Now that we have a better solution that doesn't require
monkeypatching so much Python, this is no longer needed.