Made admin user more like a normal user
Based on feedback from James, the login flow for an admin user is now
the same as a normal user. They login to the `index`, and clicking the
SecureDrop logo takes them back to the `index`. They now have an "admin"
link in the top right corner that takes them to the `admin_index`.