Always use different session cookie names for each interface
Previously, we only did this if `SECUREDROP_ENV == dev`. In production,
this doesn't cause problems because the interfaces are served to
different .onion addresses, so there is no origin collision. In
development, when both servers are running on localhost, cookies of the
same name collide (and overwrite each other) because cookies do not
consider port to be part of the origin.

Now that we have a wider variety of deployment environments, it is no
longer possible to say that we can get away with only changing the names
for the development environment. staging, for example, also provides
access to the web interfaces via localhost (port forwarded from
Vagrant), and has the same problem with name collision.

Since there is no downside to using unique names for each interface, we
just do it for all environments from now on.