Only select and de-select visible checkboxes
This avoids selecting checkboxes that are not visible due to filtering
by journalist designation. There was a hack to avoid accidentally
deleting collections that were selected, this fix replaces that with a
much cleaner and simpler solution which also prevents confusion by
seeing previously hidden items checked when clearing the filter text.