Load the JS warning bubble in the initial HTML
Loading the JS warning bubble in the initial HTML (it is hidden at first
by the CSS) triggers the loads for the icons right away. Previously, we
were generating the bubble's HTML in JS and then appending it to body,
which triggered the image loads once the element added (and thus,
visible). This caused the bubble to load without the icons at first due
to Tor's general sluggishness.

Doing it this way means the icons will probably already be loaded by the
time the user has read the JS warning and decided to click the link,
which provides a better experience overall.