Replace reply confirmation page with flashed notification
Previously, when a journalist submitted a reply to a source on the
Document Interface, they would see a confirmation page that:

1. Confirmed that the reply had been stored
2. Explained that the reply:
   1. would show up in the list of documents
   2. might eventually be deleted by the source, which could be
      indicative that it was read (although that is a bit fuzzy).

The journalist then had to click the "Return to the list of documents
for foobar..." link to return to the list of documents.

This extra step to the workflow quickly got annoying. I replaced the
reply confirmation page with a flashed notification confirming that the
reply was stored, and a redirect back to the page for the source that
was replied to.

Moved the explanation of reply/source interaction to the Journalist
User Manual and deleted the screenshot of the removed `reply.html` page.