Remove unnecessary check for gpg2 vs gpg
Now that we use Ansible and control the package dependencies via Debian
packaging, it's not possible to install SecureDrop without also
installing gpg2. Therefore, this test is no longer needed.

In addition, the response to not finding gpg2 was questionable (not sure
how the WSGI/web server would react to the assertion). This is also a
bad separation of concerns - dependency management should be separate
from the application code as much as possible.

Additionally, the initial call out to the gpg binary to check the
version was temporarily creating a key directory in `/var/www/.gnupg`,
which caused AppArmor errors. Since the version check isn't needed
anyway, the easiest way to fix this error is just to remove the check.