Fix lag in updating UI after deleting replies
Since the changes in f9969001 simplified and reduced the number of steps
in the submission deletion flow, it exposed a previously hidden problem
with deleting replies. Submissions and replies are both securely deleted
with `srm` to reduce the possibility of forensic recovery. `srm` is run
asynchronously via a job queue, since it can take a long time to run
(especially for large files).

Since Submissions were recorded in the database, we used that to
determine if they had been deleted. The files might still be on the
disk, but they would be in the job queue either being deleted or waiting
to be deleted. Replies never had a database table, so we previously
tested for whether they'd been deleted by checking for their presence on
the filesystem. This assumption is rendered invalid by the async job
queue, and the problem was made readily visible by the reduced wasted
time in the new and improved deletion UX flow. Now it is possible for a
user to select multiple documents and delete them, only to see the
replies still in the list upon being automatically redirected to the
collection page. Waiting and then refreshing will cause the replies to
disappear, showing that they had been scheduled for deletion but had not
yet been deleted when the collection list was refreshed.

The best solution to this problem was to add a similar table for Replies
to the database. This has the benefit of simplifying a lot of code,
which previously handled replies and submissions differently and can now
handle it identically or at least similarly. It may also be beneficial
in the future, as it is now easier to attach metadata to replies and
build more sophisticated data relations between them and other entities
in the system.