Fix SQLAlchemy queries for unread submissions
This regression was caused by the PEP8 fixes in cd6f4c3, which among
other things changed instances of `x == False` to `x is False`. While
this would not cause a functional change in normal Python, it does cause
a functional change in the context of SQLAlchemy's `filter*`
methods. These methods override `==` but do not appear to override `is`,
so replacing `==` with `is` breaks the query.

I have grep'ed the code and checked the diff of the breaking commit to
make sure this commit fixes all instances of this error.

Fixes #969