`./manage.py run` for development
This commit replaces `./manage.py {start,stop,restart}` with a single
command, `./manage.py run`. This command runs *both* of the interfaces
with Flask development servers, which provide useful debugging and
auto-reloading features. The logging output from both interfaces is
aggregated into a single stream, labeled for clarity, and sent to stdout
for the developer to read.

The `run` command carefully handles the case where a change to a
monitored file cannot be interpreted (e.g. if it introduces a
SyntaxError), which causes the auto-reloader to exit. This case would
break `./manage.py start` because one of the processes it had spawned
had died behind its back. This made it difficult (especially for new
developers) to understand and recover from a somewhat common error
condition. `run` does not break when this happens, but prints out a
clear message and exits both servers to avoid confusion.

Updates development documentation to recommend using `run`, and removes
references to the removed `{start,stop,restart}` commands.