Maximum codename length to prevent DoS
This is a corollary to the previous commit. While it was not mentioned
in the iSec report, it follows that if we apply a mitigation to the
journalist login to prevent DoS via submitting excessively long
passwords, we should do the same for the Source Interface - which, if
anything, is *more* vulnerable because it is publicly available as a
normal hidden service, not an authenticated hidden service like the
Document Interface.

The MAX_CODENAME_LEN is arbitrary, but is already far longer than the
longest possible codename given the current wordlist and maximum
codename length (10 words). I have tested to confirm that 128-character
codenames cannot trigger DoS.

I also modified the source logging to not print out the codename when
`valid_codename` returns False. In the context of this issue, this is
useful because an attacker who is sending enormously long codenames as a
DoS attempt would have those codenames printed to the log file, which
would make it extremely dificult to work with the logs. In general, this
is defense in depth - while logging for the Source Interface is disabled
in production, if it were turned on temporarily or accidentally, it
would be good to avoid logging all invalid codenames, as I would expect
that many of them would be *very* close to actual source's codenames
(except for one or two typos).

Adds a unit test to make sure `hash_codename` is not called if the
codename is over the maximum length, which prevents the DoS.

Corollary to resolving iSEC-15FTC-6.